fix: me, meal tanstack query migration
