+++ apps/web/src/apis/chat/_hooks/getChat.hook.ts
@@ -1,25 +1,20 @@
-import { useQueryClient } from 'react-query';
-
-import { enqueueSnackbar } from 'notistack';
+import { useQueryClient } from '@tanstack/react-query';
 
 import { getChats } from '..';
 import { useQuery } from '../../../config/react-query/useQuery';
 import { QUERY_KEYS } from '../../../constants/queryKeys';
 import { useFlow } from '../../../layouts/stackflow';
 
-
 export const useGetChats = () => {
   const { push } = useFlow();
 
   return useQuery({
-    queryKey: QUERY_KEYS.CHATS,
+    queryKey: [QUERY_KEYS.CHATS],
     queryFn: () => getChats(),
-    options: {
-      onError: () => {
-        enqueueSnackbar('채팅내역을 불러오는데 실패했습니다. 잠시 뒤 다시 시도해주세요', { variant: 'warning' });
-        push('MainActivity', {});
-      },
-    },
+    // onError: () => {
+    //   enqueueSnackbar('채팅내역을 불러오는데 실패했습니다. 잠시 뒤 다시 시도해주세요', { variant: 'warning' });
+    //   push('MainActivity', {});
+    // },
   });
 };
 
@@ -27,7 +22,7 @@ export const usePrefetchChat = async () => {
   const queryClient = useQueryClient();
 
   await queryClient.prefetchQuery({
-    queryKey: QUERY_KEYS.CHATS,
+    queryKey: [QUERY_KEYS.CHATS],
     queryFn: () => getChats(),
   });
 };
