+++ apps/web/src/apis/meal/_hooks/getMeal.ts
@@ -1,21 +1,14 @@
 import { getMealByDate } from '..';
-import useSaveTodayNutrient from '../../../activities/MainActivity/_hooks/useSaveTodayNutrient';
 import { useQuery } from '../../../config/react-query/useQuery';
 import { QUERY_KEYS } from '../../../constants/queryKeys';
-import { useUser } from '../../../recoil/auth';
+// import { useAuthUser } from '../../../providers';
 
-export const useGetMealByDate = (date: string) => {
-  const user = useUser();
-  const { saveNutrient } = useSaveTodayNutrient();
+export const useGetMealByDateQuery = (date: string) => {
+  // const user = useAuthUser();
 
   return useQuery({
     queryKey: [QUERY_KEYS.MEAL_BY_DATE, date],
     queryFn: () => getMealByDate(date),
-    options: {
-      enabled: !!user,
-      onSuccess: data => {
-        saveNutrient(data);
-      },
-    },
+    // enabled: !!user,
   });
 };
