+++ apps/web/src/providers/AuthProvider.tsx
@@ -1,6 +1,10 @@
-import { createContext, useState, useContext } from 'react';
+import { createContext, useState, useContext, useEffect } from 'react';
 
 import { User } from '@repo/ui';
+import { enqueueSnackbar } from 'notistack';
+
+import { useGetUserInfo } from '../apis/auth/_hooks/me';
+import { useFlow } from '../layouts/stackflow';
 
 type ProviderContextType = {
   user: User | null;
@@ -20,7 +24,18 @@ export function useAuthUser() {
 }
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
+  const { push } = useFlow();
   const [user, setUser] = useState<User | null>(null);
+  const { data: userData } = useGetUserInfo();
+
+  useEffect(() => {
+    if (userData) {
+      setUser(userData);
+    } else {
+      enqueueSnackbar('비타에 로그인 해 주세요!', { variant: 'warning' });
+      push('AuthActivity', {});
+    }
+  }, [userData]);
 
   return <ProviderContext.Provider value={{ user, setUser }}>{children}</ProviderContext.Provider>;
 }
